Move the route constraint classes into app/constraints
Constraints::RssFormat had no callers, so it goes rather than moves.
