+++ config/routes.rb
@@ -1,42 +1,5 @@
 require "sidekiq/web"
 
-class AdminConstraint
-  def matches?(request)
-    return false unless request.session[:user_id]
-    user = User.kept.find_by(id: request.session[:user_id])
-    user&.admin?
-  end
-end
-
-module DomainConstraints
-  def self.default_domain?(request)
-    if Rails.env.test?
-      [ "localhost", "lvh.me", "example.com" ].include?(request.host)
-    else
-      request.host == Rails.application.config.x.domain
-    end
-  end
-
-  def self.api_domain?(request)
-    if Rails.env.test?
-      request.host == "api.example.com"
-    elsif Rails.env.production?
-      request.host == "api.#{Rails.application.config.x.domain}"
-    else
-      request.host.start_with?("api.")
-    end
-  end
-end
-
-
-module Constraints
-  class RssFormat
-    def matches?(request)
-      request.format.rss?
-    end
-  end
-end
-
 Rails.application.routes.draw do
   # Helper method for subdomain redirects
   def subdomain_redirect(path = "/")
